+++ _examples/routing/writing-a-middleware/share-services/main_test.go
@@ -1,0 +1,14 @@
+package main
+
+import (
+	"testing"
+
+	"github.com/kataras/iris/v12/httptest"
+)
+
+func TestShareServices(t *testing.T) {
+	app := newApp()
+	e := httptest.New(t, app)
+
+	e.GET("/").Expect().Status(httptest.StatusOK).Body().Equal("Hello, Gophers!")
+}